% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_class_methods}
\alias{generate_class_methods}
\title{Provide a list of class methods.}
\usage{
generate_class_methods(class_methods, class_name, instance_methods, field_names,
  return_all)
}
\arguments{
\item{class_methods}{list. A named list of functions corresponding to
methods inherited from the inheritance chain.}

\item{class_name}{character. The name of the class these methods will
belong to.}

\item{instance_methods}{list. Additional methods provided by the
\code{methods} parameter to \code{setRefClass}.}

\item{field_names}{character. The names of the fields in the class.
Used to verify fields are not getting incorrectly assigned with
local (\code{<-}) instead of global (\code{<<-}) assignment.}

\item{return_all}{logical. Whether or not to return all methods.
# TODO: (RK) What does this mean?}
}
\value{
a named list.
}
\description{
Provide a list of class methods.
}

