% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_all_ref_classes}
\alias{get_all_ref_classes}
\title{Get all reference superclass names from the inheritance chain.}
\usage{
get_all_ref_classes(superClassDefs)
}
\arguments{
\item{superClassDefs}{list. The list of respective class definitions.}
}
\value{
a full character vector of superclasses in the inheritance chain,
  obtained by grabbing the slot \code{refSuperClasses} on the class definitions.
}
\description{
Given a list of reference class definitions, fetch all class and superclass
definitions from the inheritance chain.
}

