% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_all_ref_superclasses}
\alias{get_all_ref_superclasses}
\title{Get all reference superclass definitions from the inheritance chain.}
\usage{
get_all_ref_superclasses(superClassDefs)
}
\arguments{
\item{superClassDefs}{list. The list of respective class definitions.}
}
\value{
a full character vector of superclasses in the inheritance chain,
  obtained by calling \code{refSuperClasses} on the class definitions.
}
\description{
Given a list of reference class definitions, fetch all superclass
definitions from the inheritance chain.
}

