% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_superclasses_information}
\alias{get_superclasses_information}
\title{Get superclass information.}
\usage{
get_superclasses_information(contains, where)
}
\arguments{
\item{contains}{character. A character vector of super class names.}

\item{where}{environment. The environment in which to look for superClasses.
# TODO: (RK) Provide ability to look for super classes in multiple environments?}
}
\value{
a list with keys \code{superClasses}, \code{isRefSuperClass}
 indicating a character vector of super class names, whether
 each is a defined reference class or not, and the super class definitions,
 respectively.
}
\description{
Provide a list containing \code{superClasses}, (the actual class names)
\code{isRefSuperClass}, a vector of logicals indicating
whether each superclass is a reference class,
and \code{superClassDefs}, containing the actual superclass
definitions.
}

