% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inferProperties}
\alias{inferProperties}
\title{Infer the properties of the fields of a reference class.}
\usage{
inferProperties(props, what)
}
\arguments{
\item{props}{}

\item{what}{}
}
\description{
# TODO: (RK) Figure out what inferring means...
}

