% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inject}
\alias{inject}
\title{Inject a list into the current environment.}
\usage{
inject(values, where = parent.frame())
}
\arguments{
\item{values}{list. A named list of values.}

\item{where}{environment. Where should we inject this list? The
default is \code{parent.frame()}, the calling environment.}
}
\value{
\code{NULL} -- the values will be injected into \code{where}.
}
\description{
For example, if we call this function with \code{list(a = 1, b = 2)},
it will create local variables \code{a} and \code{b} with
values \code{1} and \code{2}, respectively.
}
\examples{
inject(list(a = 1, b = 2))
stopifnot(exists('a') && exists('b'))
}

