% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_active_binding}
\alias{make_active_binding}
\title{Make an active binding for a field.}
\usage{
make_active_binding(field)
}
\arguments{
\item{field}{activeBindingFunction or character.}
}
\value{
\code{activeBindingFunction}
}
\description{
Make an active binding for a field.
}

