% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_default_binding}
\alias{make_default_binding}
\title{Create a default binding (prototype) for a given field in a given environment.}
\usage{
make_default_binding(field_name, field_class, read_only = FALSE, where)
}
\arguments{
\item{field_name}{character. The name of the field.}

\item{field_class}{character. The class of the field.}

\item{read_only}{logical. Whether or not this field should be read-only.
The default is \code{FALSE}.}

\item{where}{environment. The environment in which to create the field binding.}
}
\value{
a list consisting of a \code{defaultBindingFunction} and a
   \code{uninitializedField} if the \code{field_class} refers to a virtual
   class or \code{new(field_class)} otherwise.
}
\description{
Before fields are initialized in reference classes, they should be initialized
to some value. If this is not done in the constructor, it is the responsibility
of this function to give a default value.
}

