% Generated by roxygen2 (4.0.1): do not edit by hand
\name{names_are_unique_and_non_empty}
\alias{names_are_unique_and_non_empty}
\title{Verify that the names of the given object are non-empty and distinct.}
\usage{
names_are_unique_and_non_empty(obj, what = "", error = FALSE)
}
\arguments{
\item{obj}{any R object that has \code{names}.}

\item{what}{character. A description to use when \code{error = TRUE}.
For example, if \code{what = "field"} and no names were supplied,
the error \code{"No field names supplied"} will trigger. The default
is \code{""}.}

\item{error}{logical. Whether or not to error if the names of the \code{obj}
are not unique and non-empty. The default is \code{FALSE}, in which case
a \code{logical} of length will be returned.}
}
\value{
If \code{error = TRUE}, an error will be thrown if all the names
  of the \code{obj} are not non-empty and unique. Otherwise, \code{TRUE}
  or \code{FALSE} will be returned according as the names supplied are
  unique and non-empty.
}
\description{
Since R supports duplicate list names or empty list names, this
function is handy for any verification of hash-like names: that is,
none must be empty, and they must all be distinct.
}
\examples{
stopifnot(names_are_unique_and_non_empty(c()))
stopifnot(names_are_unique_and_non_empty(list(a = 1, b = 2))
stopifnot(!names_are_unique_and_non_empty(c("")))
stopifnot(!names_are_unique_and_non_empty(list(a = 1, a = 2))
}

