% Generated by roxygen2 (4.0.1): do not edit by hand
\name{process_field_information}
\alias{process_field_information}
\title{Determine class and prototypes for fields from super class hierarchy.}
\usage{
process_field_information(fieldNames, fieldClasses, fieldPrototypes,
  superClassDefs)
}
\arguments{
\item{fieldClasses}{character. The names for each field.}

\item{fieldClasses}{character. The classes for each field.}

\item{fieldClasses}{list or function. The prototype for each field
(i.e. default binding).}

\item{superClassDefs}{list. A list of super class definitions (things of
class \code{refClassRepresentation}, obtained using \code{methods::getClass}.}
}
\description{
If \code{superClassDefs} is a list of parent classes, their definitions will
be traversed in reverse order (so that nearer classes respect the inheritance
chain) and the according field and prototypes extracted.
}

