% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{refclass}
\alias{refclass}
\alias{refclass-package}
\title{R reference classes.}
\description{
This package is an attempt to refactor R's built-in reference class
code to make it more understandable, documented, and tested.
}

