% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_dummy_field}
\alias{set_dummy_field}
\title{Set a dummy field.}
\usage{
set_dummy_field(self, meta_name, field_class, field_name, only_once, value)
}
\arguments{
\item{self}{environment. The environment of the reference class objet.}

\item{meta_name}{character. The name to bind the dummy field to.}

\item{field_class}{character. The class of the field.}

\item{field_name}{character. The name of the field.}

\item{only_once}{logical. Whether or not to set the dummy field only once.}

\item{value}{The R object to bind to the dummy \code{field_name}.}
}
\value{
No return value -- the \code{self} environment gets modified to
  contain a \code{meta_name} value.
}
\description{
Set a dummy field.
}

