% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ref_class_information}
\alias{ref_class_information}
\alias{setRefClass}
\title{Fetch all necessary information to construct a reference class generator function.}
\usage{
ref_class_information(Class, contains, fields, refMethods, where)

setRefClass(Class, fields = character(), contains = character(),
  methods = list(), where = topenv(parent.frame()), ...)
}
\arguments{
\item{refMethods}{list. A named list of reference class methods.}

\item{Class}{character. The name of the newly created reference class.}

\item{fields}{character. The field for the reference class. Fields are
equivalent to "members" in other languages, and can represent any data.
In this particular case, the convention is something like
\code{list(id = "numeric", data = "data.frame")} to specify two fields,
\code{id} and \code{data} with required types \code{numeric} and
\code{data.frame}. The default is \code{character()}.}

\item{contains}{character. The parent reference classes of this reference
class. The default is \code{character()}. # TODO: (RK) Document this behavior better.}

\item{methods}{list. The methods for instances of this reference class.
In particular, the names of this list represent the names of the methods,
and the values must be functions. For example,
 \code{list(foo = function(x) print(x)} would define a method \code{foo}
that we could call with \code{reference_class_object$foo(...)}.
The default is to default is to define no methods (\code{list()}).}

\item{where}{environment. Where to store the definition of the reference
class. The default is \code{topenv(parent.frame())}.}

\item{...}{additional arguments to the \code{setClass} function call
that gets called under the hood. (Note that a reference class generator
is an extension of a classGeneratorFunction in the S3 sense.)}
}
\value{
the fully formed reference class object generator. For example,
  if we call \code{x <- setRefClass(...)}, then we can call \code{x$new(...)}
  to invoke the constructor for the object \code{x}. The class of the
  reference class object generator is \code{"refObjectGenerator"}.
# TODO: (RK) More documentation and examples.
}
\description{
Fetch all necessary information to construct a reference class generator function.

Reference class objects are the closest approximation of the concepts of
object-oriented programming available in R. While S3 and S4 classes serve
the useful purpose of providing dispatching mechanisms on generic functions,
reference classes allow for mutability of data on the underlying object
regardless of calling convention (or nasty syntax like replace methods).
}
\details{
# TODO: (RK) More description here.
}

