% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simple_error}
\alias{simple_error}
\title{Call base::stop with interpolated arguments.}
\usage{
simple_error(message, ..., call. = TRUE)
}
\arguments{
\item{message}{character. The message to call \code{base::stop} with.}

\item{...}{any instances of \code{"\%s"} in \code{message} will be replaced
with the respective additional arguments.}

\item{call.}{logical. Whether or not to show a stack trace. The default is
\code{TRUE}.}
}
\description{
Call base::stop with interpolated arguments.
}
\examples{
stopifnot(identical('hello world', tryCatch(error = function(e) e$message,
  simple_error("hello \%s", "world"))))
}

