% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to_env}
\alias{to_env}
\title{Convert a list to an environment.}
\usage{
to_env(x)
}
\arguments{
\item{x}{list. The list to convert to an environment.}
}
\value{
an environment with keys coming from the names of \code{x}
  and values coming from the respective values of \code{x}.
}
\description{
If the list if of length 0, be careful to not use \code{as.environment}
as it will break.
}

