% Generated by roxygen2 (4.0.1): do not edit by hand
\name{format_types_list}
\alias{format_types_list}
\title{Format a list or character vector of types or error if invalid.}
\usage{
format_types_list(props, what)
}
\arguments{
\item{props}{ANY}

\item{what}{ANY}
}
\description{
A reference class must specify its fields, for example,
  \code{c(data = "data.frame", ntrees = "integer")}. These fields
can be provided in several formats. The goal of this function is to
ensure that format is correct, make small modifications if necessary,
and error if there are problems.
}
\details{
The following are valid formats:

\items{
}

# TODO: (RK) Figure out what inferring means...
}

