% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_fields}
\alias{parse_field}
\alias{parse_fields}
\title{Parse out field class names and prototypes.}
\usage{
parse_fields(fields, where)

parse_field(field_name, field_value, where)
}
\arguments{
\item{fields}{list. A named list of \code{character}s or \code{function}s.
If \code{character}, this represents the class name of the field
(for example, \code{"data.frame"}). If a function, an active binding
that will be used for the field prototype. For example, if we have
a field \code{radius} and a field \code{diameter}, we could set the
default value of \code{radius} (its prototype) to
\code{function() 2 * radius}, which wil be evaluated within the context
of the reference class environment.}

\item{where}{environment. Where to look for class definitions.}

\item{field_name}{character. The name of the field. (This is primarily
used for error messages when the \code{field_value} does not validate
all conditions.)}

\item{field_value}{character or function. See \code{fields}.}
}
\value{
a list with three keys, \code{names}, \code{classes} and \code{prototypes} of
   class names and prototypes, respectively. Prototypes are just the
   default values of the respective fields.
}
\description{
Parse out field class names and prototypes.

A field definition (\code{field_value}) can be either a character or
a binding function (see the \code{fields} parameter).
}

